Patch 1.3.1 changes!
Added the second option, 2 Players!

Changed the "Class ManualPlayer" to -> "Class FirstPlayer".
Added a "Class SecondPlayer".
Edited the "Class Highscore" - it can now stream-write for both single and multi player options.
Edited the "Class UserInterface" - it can now function with the 2 players option.
Edited the "Class GameManager" - Made a major changing in this class. Added a few more functions, moved some statements here and there, so the game can identify and function if the user has chosen to play the multi players option.